Final_Course_Project/FinalCourseProject/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
catboost>=1.2.8
pandas>=2.2.0
numpy>=1.24.0
scikit-learn>=1.4.0
python-dotenv==1.0.0
Werkzeug==3.0.0
